requirements.txt
APScheduler==3.6.1
Babel==2.7.0
certifi==2019.6.16
cloudinary==1.16.0
dj-static==0.0.6
Django==2.2.2
django-apscheduler==0.3.0
django-cors-headers==3.0.2
django-extra-fields==1.2.3
django-filter==2.1.0
django-phone-field==1.7.1
django-phonenumber-field==3.0.1
django-ratings==0.3.7
djangorestframework==3.9.4
gunicorn==19.9.0
Markdown==3.1.1
meld3==1.0.2
mock==3.0.5
mysqlclient==1.4.2.post1
phonenumbers==8.10.14
Pillow==6.1.0
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
supervisor==4.0.3
tzlocal==2.0.0
urllib3==1.25.3
yapf==0.27.0
